+++ op/upfirdn2d.py
@@ -1,6 +1,7 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 
+import functools
 import os
 
 import torch
@@ -9,13 +10,18 @@ from torch.nn import functional as F
 from torch.utils.cpp_extension import load
 
 module_path = os.path.dirname(__file__)
-upfirdn2d_op = load(
-    "upfirdn2d",
-    sources=[
-        os.path.join(module_path, "upfirdn2d.cpp"),
-        os.path.join(module_path, "upfirdn2d_kernel.cu"),
-    ],
-)
+
+
+@functools.cache
+def _load_upfirdn2d():
+    """JIT-build the CUDA extension on first use, not at import. See op/fused_act.py."""
+    return load(
+        "upfirdn2d",
+        sources=[
+            os.path.join(module_path, "upfirdn2d.cpp"),
+            os.path.join(module_path, "upfirdn2d_kernel.cu"),
+        ],
+    )
 
 
 class UpFirDn2dBackward(Function):
@@ -30,7 +36,7 @@ class UpFirDn2dBackward(Function):
 
         grad_output = grad_output.reshape(-1, out_size[0], out_size[1], 1)
 
-        grad_input = upfirdn2d_op.upfirdn2d(
+        grad_input = _load_upfirdn2d().upfirdn2d(
             grad_output,
             grad_kernel,
             down_x,
@@ -67,7 +73,7 @@ class UpFirDn2dBackward(Function):
 
         gradgrad_input = gradgrad_input.reshape(-1, ctx.in_size[2], ctx.in_size[3], 1)
 
-        gradgrad_out = upfirdn2d_op.upfirdn2d(
+        gradgrad_out = _load_upfirdn2d().upfirdn2d(
             gradgrad_input,
             kernel,
             ctx.up_x,
@@ -116,7 +122,7 @@ class UpFirDn2d(Function):
 
         ctx.g_pad = (g_pad_x0, g_pad_x1, g_pad_y0, g_pad_y1)
 
-        out = upfirdn2d_op.upfirdn2d(
+        out = _load_upfirdn2d().upfirdn2d(
             input, kernel, up_x, up_y, down_x, down_y, pad_x0, pad_x1, pad_y0, pad_y1
         )
         out = out.view(-1, channel, out_h, out_w)